New step and page for Event Search
